Zoom control
Allow a keyboard operator to control the zoom level of the camera.

Next it'd be great to do better feature detection and to auto-focus based on orb size. If orbs are larger, zoom until they get smaller. Track the maximum size stop zooming when a minima is reached. Obviously zoom in/out if the minima is passed.